+++ react_frontend/package.json
@@ -48,11 +48,11 @@
     "react-google-login": "^3.0.10",
     "react-popout": "^1.0.1",
     "react-redux": "^5.0.7",
-    "react-relay": "^1.4.1",
+    "react-relay": "^1.5.0",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^4.0.8",
     "redux": "^3.7.2",
-    "relay-runtime": "^1.4.1",
+    "relay-runtime": "^1.5.0",
     "styled-components": "^3.1.6"
   }
 }
